Added Spline Path to Art Pass Map
I've Changed the Name of the Art Pass to Level 1 Art Pass 1. In that map I've added a spline that uses the newly updated Spline Tech with the camera.

I've also updated the Lighting for the Level 1 Art Pass 1 Map.

I've added a few more decals to the roads (Subject to change)

Fixed a few Collisions on some assets.
Reworked the Jump Pad.