Update respawn, drone and projectiles, ragdoll sys. Start working on Car's ragdoll
Respawn is completely based on the location of the spline points now, so the spline has to be above the 'ground', working as a real path.
Little tweak in BP to find closest spline point, it  is more effective now.
BP CeilingLIght affects performance heavily, it is set to NOT Affect World at the moment.