Tech update: see description
Using the right character with right animation now, relative location after respawn is reset, which offset issue (during 21/5 playtest) is not there too.
Scaled the destructible mesh of the boxes back to normal size.
Scaled the boxes inside the level up.
New Column in DT_DroneProjectile? for the sound effect.