End Roof Section Meshing & More
I've Almost Finished Meshing the Rooftops.
I also Added a car track the the end Section of the Level.
I Added Coins to the first half of the Level and some section of the Roof top.
I've added The Bonus Coin the Level.
I've Adjusted the Spline Path at the Transition from the End roof to the last section.
Fixed Floating Meshes.
Fixed Z-fighting.

That is about it for the Changes, I did some Minor Tweaks to the Car Track splines and adjusted more Props.